Addition of maxlines parameter to function csv_file_encoding in dwcs_utils.py
The parameter, if supplied, will stop the function from attempting to determine encoding after maxlines tries.
if the parameter is not supplied, or is 0, there will be no stop and potitally all lines will be read.

function represents_int was added to the file to error check the maxlines paramter incase strings and such were supplied instead of a integeer parameter.

Test cases were added, both for the new function represents_int and to test that file encoding does stop after x number of maxlines. The maxline stepers were added to dwca_utils_test.py:test_csv_file_encoding towards the end of the function