Extract occurrence core actor reorders header causing errors
The list of fields in dwca_core_to_tsv is reordered in the header and the data row remains in the original order. For archives with a different header order, the data columns don't match the reordered header in the extracted file output. Keeping headers in the same order instead