Geography cleaner termlist isn't constructed from cleaned header
The cleantermlist is a list of single value lists however the cleaned header contains a list of strings. When checking if an element in cleantermlist is present in the cleaned header, no match is found (for example ['countrycode'] != 'countrycode'. Modified cleantermlist to be a list of strings as well so that it matches the cleanedheader.